refactor(svgprops): move svg props to global types
