--- src/assets/svg/style.d.ts
@@ -1,6 +1,0 @@
-export interface SVGProps {
-  width?: number | string
-  height?: number | string
-  fill?: string
-  stroke?: boolean
-}
